feat: translated index server

+++ src/bot/index.ts
@@ -18,15 +18,14 @@ import { PrismaClient } from "@prisma/client";
 import { VoiceUpdate } from "./actions/VoiceUpdate";
 import { FinishedEvent } from "./actions/FinishedEvent";
 import { DeleteEvent } from "./actions/DeleteEvent";
-import { Seller } from "./commands/Seller";
 import { handleGuildEvents } from "./handlers/guildEventHandler";
 import { ProcessDeposit } from "./actions/ProcessDeposit";
-import { Admin } from "./commands/Admin";
-import { Event } from "./commands/Event";
-import { Global } from "./commands/Global";
 import "./data/itemsLoader";
 import { initI18n } from "./locales/initI18n";
 import { isInCooldown, setCooldown } from "./utils/cooldown";
+import { useT } from "./utils/useT";
+import { handleCommandInteraction } from "./handlers/handleCommandInteraction";
+import { handleButtonInteraction } from "./handlers/handleButtonInteraction";
 
 const client = new Client({
   intents: [
@@ -53,158 +52,14 @@ client.once("ready", async () => {
 });
 
 client.on("interactionCreate", async (interaction) => {
-  const member = await interaction?.guild?.members.fetch(interaction.user.id);
-  const isAdmin = member?.permissions.has(PermissionsBitField.Flags.Administrator);
-
+  //comandos de barra
   if (interaction.isCommand()) {
-    const { commandName } = interaction;
-    try {
-      //comandos a serem usados apenas por administrador para configurações
-      if (isAdmin) {
-        const wasCommandExecuted = await Admin({
-          commandName,
-          interaction,
-          prisma,
-        });
-
-        if (wasCommandExecuted) return;
-      } else {
-        // Se não for admin e tentar usar um comando restrito, retorna uma mensagem
-        const adminCommands = [
-          "setup",
-          "atualizar-taxa-guild",
-          "atualizar-taxa-vendedor",
-          "depositar-guild",
-          "pagar-membro",
-          "sacar-guild",
-          "confiscar-saldo",
-          "remover-bot",
-        ];
-        if (adminCommands.includes(commandName)) {
-          return await interaction.reply("Apenas um **Administrador** pode usar esse comando!");
-        }
-      }
-
-      //comandos qualquer cargo consegue usar
-      const wasCommandExecuted = await Global({
-        commandName,
-        interaction,
-        member,
-        prisma,
-      });
-
-      if (wasCommandExecuted) return;
-
-      //comandos para evento
-      const event = await prisma.event.findFirst({
-        where: {
-          channelID: interaction.channelId,
-        },
-      });
-
-      //comandos só podem ser usados caso exista um evento no canal que foi usado o comando
-      if (event) {
-        if (event.status === "closed") {
-          await interaction.reply("Evento já fechado, não é possivel mais usar comandos!");
-          return;
-        }
-
-        const guildData = await prisma.guilds.findUnique({
-          where: {
-            guildID: interaction.guild?.id,
-          },
-        });
-
-        if (!guildData) {
-          await interaction.reply("Guild não encontrada no banco de dados!");
-          return;
-        }
-
-        const isManager = await interaction?.guild?.roles.fetch(guildData?.eventManagerRoleID ?? "");
-
-        if (commandName === "vendedor") {
-          if (!isManager && !isAdmin) {
-            return interaction.reply("Apenas um Manager ou Administrador pode adicionar um vendedor ao evento!");
-          }
-
-          await Seller({
-            prisma,
-            interaction,
-            event,
-          });
-          return;
-        }
-
-        if (!event.seller) {
-          await interaction.reply(`Adicione um vendedor antes de usar /${commandName}!`);
-          return;
-        }
-
-        const isSeller = event.seller === interaction.user.id; // Verifica se é o vendedor do evento
-
-        if (isSeller || isAdmin || isManager) {
-          return await Event({
-            commandName,
-            event,
-            interaction,
-            prisma,
-            guildData,
-          });
-        } else {
-          return await interaction.reply(`\n\`/${commandName}\` só pode ser usado pelo vendedor!`);
-        }
-      } else {
-        return await interaction.reply(`\n\`/${commandName}\` só pode ser usado em um canal de eventos!`);
-      }
-    } catch (error) {
-      console.error("Erro ao processar interação:", error);
-      if (!interaction.replied && !interaction.deferred) {
-        await interaction.reply(`Ocorreu um erro ao processar o comando: ${commandName}`);
-      } else {
-        console.error(`Falha ao processar comando: ${commandName}`);
-      }
-    }
+    await handleCommandInteraction(interaction, prisma);
   }
 
-  //BOTÃO DE CRIAR EVENTO
+  //comandos de botão
   if (interaction.isButton()) {
-    if (!interaction.guild) return; // Garante que estamos em um servidor
-    //aqui tambem fazer verificação quando o cargo de criar de eventos for criado
-
-    try {
-      const guildData = await prisma.guilds.findUnique({
-        where: {
-          guildID: interaction.guild?.id,
-        },
-      });
-
-      // const isManager = await interaction.guild.roles.fetch(guildData?.eventManagerRoleID ?? "");
-
-      // if (!isAdmin && !isManager) {
-      //   await interaction.deferUpdate();
-      //   return;
-      // }
-
-      if (!guildData || interaction.channelId !== guildData?.newEventChannelID) {
-        return await interaction.deferUpdate();
-      }
-
-      if (interaction.customId === "create_event") {
-        await interaction.deferUpdate();
-        await OpenEvent({ interaction, guildData });
-      }
-    } catch (error) {
-      console.error("Erro ao tentar criar evento!");
-
-      // Após deferir a interação, não é possível mais usar reply
-      if (!interaction.replied && !interaction.deferred) {
-        // Apenas usa reply se a interação não foi deferida nem respondida
-        return await interaction.reply("Erro ao tentar criar evento!");
-      } else {
-        // Caso tenha sido deferido ou já respondido, apenas loga o erro.
-        console.error("Falha ao tentar criar evento, interação já foi deferida ou respondida.");
-      }
-    }
+    await handleButtonInteraction(interaction, prisma);
   }
 });
 
@@ -239,7 +94,7 @@ client.on(
 
       if ("send" in channel && typeof channel.send === "function") {
         try {
-          await channel.send(`${user}, reação não processada porque a mensagem original foi apagada.`);
+          await channel.send(`${user} reaction message not found`);
         } catch (err) {
           console.warn("Erro ao enviar mensagem no canal:", err);
         }
@@ -250,13 +105,28 @@ client.on(
 
     if (message.author?.id !== client.user?.id) return;
 
+    const guildData = await prisma.guilds.findUnique({
+      where: {
+        guildID: message.guild?.id,
+      },
+    });
+
+    if (!guildData) {
+      console.log(
+        `configurações da guild: ${reaction.message.guild?.name}[${reaction.message.guildId}] não encontradas!`
+      );
+      return;
+    }
+
+    const t = useT(guildData.language);
+
     //verificação de tempo de reação
     if (isInCooldown(user.id)) {
       await reaction.users.remove(user.id);
       const channel = reaction.message.channel;
       if ("send" in channel && typeof channel.send === "function") {
         try {
-          const msg = await channel.send(`${user}, aguarde 2 segundos antes de reagir novamente.`);
+          const msg = await channel.send(t("index.waitSendMessage", user));
           setTimeout(() => msg.delete().catch(() => {}), 3000);
         } catch (err) {
           console.warn("Erro ao enviar mensagem de cooldown:", err);
@@ -273,12 +143,6 @@ client.on(
 
     const channelID = message.channel.id;
 
-    const guildData = await prisma.guilds.findUnique({
-      where: {
-        guildID: message.guild?.id,
-      },
-    });
-
     const financeChannelID = guildData?.financialChannelID;
     const participationChannelID = guildData?.participationChannelID;
 
@@ -291,27 +155,26 @@ client.on(
     if (channelID === financeChannelID) {
       if (reaction.emoji.name === "✅") {
         await reaction.users.remove(user.id);
-        const description = embed.description;
+        const footerText = embed.footer?.text;
 
         // Extrair o nome do evento
-        const eventNameMatch = description?.match(/(Evento\s\d+)/);
+        const eventNameMatch = footerText?.match(/ID:(Event\s\d+)/);
         const eventName = eventNameMatch ? eventNameMatch[1] : null;
 
-        // Extrair o valor total informado pelo usuário
-        const totalValueMatch = description?.match(/informou o valor total de `([\d,\.]+)`/);
+        // Extrair o valor total informado
+        const totalValueMatch = footerText?.match(/v:\s*([\d,.]+)/);
         const totalValue = totalValueMatch ? Number(totalValueMatch[1].replace(/[,.]/g, "")) : null;
-        4;
 
         if (!eventName) {
           if (message.channel && "send" in message.channel) {
-            await message.channel.send("Não foi possível identificar o evento!");
+            await message.channel.send(t("index.eventUnidentified"));
           }
           return;
         }
 
         if (!totalValue) {
           if (message.channel && "send" in message.channel) {
-            return await message.channel.send("Não foi possível identificar o valor do depósito!");
+            return await message.channel.send(t("index.depositUnidentified"));
           }
 
           return;
@@ -325,7 +188,7 @@ client.on(
 
         if (!event) {
           if (message.channel && "send" in message.channel) {
-            await message.channel.send("Evento não encontrado");
+            await message.channel.send(t("index.eventNotFound"));
           }
 
           return;
@@ -371,11 +234,16 @@ client.on(
 
             if (confirmationMessage) {
               const updatedEmbed = new EmbedBuilder()
-                .setTitle(`Depósito do ${event.eventName} Confirmado`)
+                .setTitle(
+                  t("index.confirmedDepositEmbed.title", {
+                    eventName: event.eventName,
+                  })
+                )
                 .setDescription(
-                  `<@${user.id}> informou o valor de **${totalValue.toLocaleString(
-                    "en-US"
-                  )} de pratas**, e ele já foi depositado no saldo dos participantes do **${event.eventName}**.`
+                  t("index.confirmedDepositEmbed.description", {
+                    userId: user.id,
+                    totalValue: totalValue.toLocaleString("en-US"),
+                  })
                 )
                 .setColor("Green");
 
@@ -394,9 +262,9 @@ client.on(
       const creatorNameMatch = embed.title?.match(/Criado por ([^-]+)/);
       const creatorName = creatorNameMatch ? creatorNameMatch[1].trim() : null;
 
-      const eventNumberMatch = embed.title?.match(/^Evento (\d+)\b/);
+      const eventNumberMatch = embed.title?.match(/^Event (\d+)\b/);
       const eventNumber = eventNumberMatch?.[1] || "";
-      const keyTitle = `Evento ${eventNumber}`;
+      const keyTitle = `Event ${eventNumber}`;
 
       const eventCreator = creatorName === user.username;
 
@@ -410,6 +278,7 @@ client.on(
           embed,
           keyTitle,
           eventNumber,
+          guildData,
         });
 
         return;
@@ -431,6 +300,7 @@ client.on(
           prisma,
           keyTitle,
           eventNumber,
+          guildData,
         });
 
         return;
@@ -445,6 +315,7 @@ client.on(
           prisma,
           keyTitle,
           reaction,
+          guildData,
         });
 
         return;
@@ -459,6 +330,7 @@ client.on(
           reaction,
           user,
           creatorName,
+          guildData,
         });
 
         return;
@@ -472,6 +344,16 @@ client.on("voiceStateUpdate", async (oldState, newState) => {
   // Ignorar bots
   if (oldState.member?.user.bot || newState.member?.user.bot) {
     return;
+  }
+
+  const guildData = await prisma.guilds.findUnique({
+    where: {
+      guildID: oldState.guild?.id,
+    },
+  });
+
+  if (!guildData) {
+    return console.log("Guild não encontrada, servidor", oldState.guild.name);
   }
 
   const event = await prisma.event.findFirst({
@@ -488,6 +370,7 @@ client.on("voiceStateUpdate", async (oldState, newState) => {
     newState,
     oldState,
     prisma,
+    guildData,
   });
 });
